+++ src/model2/factories/index.ts
@@ -1,2 +1,3 @@
 export * from "./account_flags_factory";
 export * from "./account_thresholds_factory";
+export * from "./signer_factory";
